+++ app/src/main/java/io/github/muntashirakon/AppManager/filters/options/FilterOptions.java
@@ -27,6 +27,7 @@ public final class FilterOptions {
             case "installer": return new InstallerOption();
             case "last_update": return new LastUpdateOption();
             case "min_sdk": return new MinSdkOption();
+            case "native_lib": return new NativeLibOption();
             case "permissions": return new PermissionsOption();
             case "pkg_name": return new PackageNameOption();
             case "running_apps": return new RunningAppsOption();